requirements.txt
asgiref==3.7.2
Django==4.2.3
django-bootstrap-modal-forms==2.2.0
django-bootstrap3==22.2
django-forms-bootstrap==3.1.0
django-widget-tweaks==1.4.12
mysqlclient==2.2.0
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.5.0
